+++ services/src/test/java/ar/edu/itba/paw/services/CommunityServiceImplTest.java
@@ -1147,7 +1147,7 @@ class CommunityServiceImplTest {
 
         @Override
         public void sendWeeklyModeratorDigest(final List<EmailRecipient> moderatorRecipients,
-                                              final int pendingRequestCount) {
+                                              final long pendingRequestCount) {
         }
 
         @Override
